requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib
pydantic-settings
pydantic[email]
bcrypt==4.0.1
flask-cors
loguru
setuptools
pydub
psutil
python-multipart
numpy
moviepy
httpx
yt-dlp
loguru
modal
ngrok
faster-whisper